backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pdfplumber==0.11.4
python-docx==1.1.2
python-multipart==0.0.9
pydantic==2.9.2
pytest==8.3.3
pytest-asyncio==0.24.0
httpx[http2]==0.27.2
openpyxl==3.1.5
